Move record logic into parent, add hotkey.
